requirements.txt
annotated-types==0.7.0
faster-whisper>=1.0.0
pytest>=7.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=23.2.0
openai>=1.30.0
anyio==4.12.1
assemblyai==0.48.4
certifi==2026.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
websockets==16.0
